move search function to EventPanel component
