+++ client/src/ListItem.jsx
@@ -7,6 +7,27 @@ const ListItem = ({ listItem, handleListItemClick }) => {
     likeClass = "fas fa-heart";
   }
 
+  let img;
+
+  if (listItem.event.logo_url) {
+    img = <img
+      className="img-fluid"
+      src={listItem.event.logo_url}
+      data-id={eventId}
+      onClick={handleListItemClick}
+      alt="no image"
+    />
+  } else {
+    img = <img
+      className="img-fluid"
+      style={{ width: "300px", height: "100%", minHeight: "5rem", backgroundColor: "#ddd" }}
+      src={listItem.event.logo_url}
+      data-id={eventId}
+      onClick={handleListItemClick}
+
+    />
+  }
+
   return (
 
     <div
@@ -16,13 +37,7 @@ const ListItem = ({ listItem, handleListItemClick }) => {
     >
       <div className="row">
         <div className="col-md-5 list-img">
-          <img
-            className="img-fluid"
-            src={listItem.event.logo_url}
-            alt="no image"
-            data-id={eventId}
-            onClick={handleListItemClick}
-          />
+          {img}
         </div>
         <div className="col-md-7 list-txt">
           <div
