+++ client/styles/navbar.scss
@@ -204,7 +204,11 @@
 
     .form-wrap {
       padding: 0.15rem;
-      padding-top: 0;
+      margin-top: 0.3rem;
+    }
+
+    label {
+      font-size: 1.2rem;
     }
   }
